% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MLB2016}
\alias{MLB2016}
\title{MLB2016.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2462 rows and 11 columns.}
\usage{
MLB2016
}
\description{
MLB data 2016.
}
\details{
All wagering lines from Pinnacle for the 2016  MLB season
}
\keyword{datasets}
