% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MLB2016}
\alias{MLB2016}
\title{MLB2016.}
\format{A data frame with 20 variables: \code{year}, \code{sex},
  \code{name}, \code{n} and \code{prop} (\code{n} divided by total number
  of applicants in that year).}
\usage{
MLB2016
}
\description{
All wagering lines from Pinnacle for the 2016  MLB season
}
\keyword{datasets}
