% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MLB2016}
\alias{MLB2016}
\title{MLB2016.}
\format{A data frame with 14 variables:
\describe{
\item{\code{EventDateTime}}{Time of the game in UTC}
\item{\code{TeamName1}}{Team name of the Away Team}
\item{\code{TeamName2}}{Team name of the Home Team}
\item{\code{StartingPitcher1}}{Starting pitcher Away Team}
\item{\code{StartingPitcher2}}{Starting pitcher Home Team}
\item{\code{GameScoreTeam1}}{Runs scored by Away Team}
\item{\code{GameScoreTeam2}}{Runs scored by Home Team}
\item{\code{EnteredDateTime}}{Time of the wager line in UTC}
\item{\code{SpreadTeam1}}{Spread Handicap for Away Team}
\item{\code{SpreadUS1}}{Spread US odds for Away Team}
\item{\code{SpreadUS2}}{Spread US odds for Home Team}
\item{\code{MoneyUS1}}{Moneyline US odds for Away Team}
\item{\code{MoneyUS2}}{Moneyline US odds for Home Team}
\item{\code{TotalPoints}}{Total runs handicap}
\item{\code{TotalUSOver}}{Total runs US odds for Over}
\item{\code{TotalUSUnder}}{Total runs US odds for Under}
}}
\usage{
MLB2016
}
\description{
MLB data 2016.
}
\details{
All wagering lines from Pinnacle for the 2016  MLB season
}
\keyword{datasets}
