% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USA_Election_2016}
\alias{USA_Election_2016}
\title{USA_Election_2016}
\format{A data frame with 12 variables:
\describe{
\item{\code{EnteredDateTime}}{Time of the wager line in UTC}
\item{\code{TeamName1}}{Team name of the Away Team}
\item{\code{TeamName2}}{Team name of the Home Team}
\item{\code{SpreadTeam1}}{Spread Handicap for Away Team}
\item{\code{SpreadUS1}}{Spread US odds for Away Team}
\item{\code{SpreadUS2}}{Spread US odds for Home Team}
\item{\code{MoneyUS1}}{Moneyline US odds for Away Team}
\item{\code{MoneyUS2}}{Moneyline US odds for Home Team}
\item{\code{TotalPoints}}{Total runs handicap}
\item{\code{TotalUSOver}}{Total runs US odds for Over}
\item{\code{TotalUSUnder}}{Total runs US odds for Under}
}}
\usage{
USA_Election_2016
}
\description{
US Presidential Election data 2016.
}
\details{
All lines from Pinnacle for the 2016 US Presidential Election
}
\keyword{datasets}
